Move folder /anim to /framework/anim
