+++ src/framework/anim/binder/anim-binder.js
