+++ src/framework/anim/constants.js
