+++ src/framework/anim/controller/anim-node.js
@@ -1,4 +1,4 @@
-import { Vec2 } from '../../core/math/vec2.js';
+import { Vec2 } from '../../../core/math/vec2.js';
 
 /**
  * AnimNodes are used to represent a single animation track in the current state. Each state can
