+++ src/framework/anim/controller/constants.js
