+++ src/framework/anim/evaluator/anim-clip.js
@@ -1,7 +1,7 @@
 import { AnimSnapshot } from './anim-snapshot.js';
 
 /** @typedef {import('./anim-track.js').AnimTrack} AnimTrack */
-/** @typedef {import('../../core/event-handler.js').EventHandler} EventHandler */
+/** @typedef {import('../../../core/event-handler.js').EventHandler} EventHandler */
 
 // TODO: add configurable looping start/end times?
