+++ src/framework/anim/evaluator/anim-curve.js
