+++ src/framework/anim/evaluator/anim-data.js
