+++ src/framework/anim/evaluator/anim-snapshot.js
