+++ src/framework/components/anim/component-layer.js
@@ -1,7 +1,7 @@
 import { Debug } from '../../../core/debug.js';
-import { AnimTrack } from '../../../anim/evaluator/anim-track.js';
-import { AnimTransition } from '../../../anim/controller/anim-transition.js';
-import { ANIM_LAYER_OVERWRITE } from '../../../anim/controller/constants.js';
+import { AnimTrack } from '../../anim/evaluator/anim-track.js';
+import { AnimTransition } from '../../anim/controller/anim-transition.js';
+import { ANIM_LAYER_OVERWRITE } from '../../anim/controller/constants.js';
 import { math } from '../../../core/math/math.js';
 
 /** @typedef {import('./component.js').AnimComponent} AnimComponent */
