+++ src/framework/components/anim/component.js
@@ -1,17 +1,17 @@
 import { Debug } from '../../../core/debug.js';
 import { Asset } from '../../asset/asset.js';
 
-import { AnimEvaluator } from '../../../anim/evaluator/anim-evaluator.js';
-import { AnimController } from '../../../anim/controller/anim-controller.js';
+import { AnimEvaluator } from '../../anim/evaluator/anim-evaluator.js';
+import { AnimController } from '../../anim/controller/anim-controller.js';
 
 import { Component } from '../component.js';
 
 import {
     ANIM_PARAMETER_BOOLEAN, ANIM_PARAMETER_FLOAT, ANIM_PARAMETER_INTEGER, ANIM_PARAMETER_TRIGGER, ANIM_CONTROL_STATES
-} from '../../../anim/controller/constants.js';
+} from '../../anim/controller/constants.js';
 import { AnimComponentBinder } from './component-binder.js';
 import { AnimComponentLayer } from './component-layer.js';
-import { AnimStateGraph } from '../../../anim/state-graph/anim-state-graph.js';
+import { AnimStateGraph } from '../../anim/state-graph/anim-state-graph.js';
 import { Entity } from '../../entity.js';
 
 /** @typedef {import('./system.js').AnimComponentSystem} AnimComponentSystem */
